# Implement itemDefaults for CompletionList