Fixes #2475