@JessicaJHee 2023-03-16
This PR currently supports `editRange` and `insertTextFormat` in `itemDefaults`. I.e., the completion list response will include something like: 

```
    "itemDefaults": {
        "editRange": {
            "insert": {
                "start": {
                    "line": 5,
                    "character": 15
                },
                "end": {
                    "line": 5,
                    "character": 15
                }
            },
            "replace": {
                "start": {
                    "line": 5,
                    "character": 15
                },
                "end": {
                    "line": 5,
                    "character": 19
                }
            }
        },
        "insertTextFormat": 2
    }
```