@rgrunber 2023-03-22
The difference looks something like this :

<details>
<summary>diff</summary>

```diff
-[Info  - 15:28:53] Mar. 22, 2023, 3:28:53 p.m. Completion request completed
-[Trace - 15:28:53] Received response 'textDocument/completion - (30)' in 386ms.
+[Info  - 15:32:22] Mar. 22, 2023, 3:32:22 p.m. Completion request completed
+[Trace - 15:32:22] Received response 'textDocument/completion - (29)' in 323ms.
 Result: {
     "isIncomplete": false,
     "items": [
@@ -8,31 +8,7 @@
             "kind": 2,
             "detail": "Completion.completion_1(String title, String text, int order) : int",
             "sortText": "999999035",
-            "insertText": "completion_1",
-            "insertTextFormat": 2,
-            "textEdit": {
-                "newText": "completion_1(${1:null}, ${2:null}, ${3:0})",
-                "insert": {
-                    "start": {
-                        "line": 20,
-                        "character": 13
-                    },
-                    "end": {
-                        "line": 20,
-                        "character": 13
-                    }
-                },
-                "replace": {
-                    "start": {
-                        "line": 20,
-                        "character": 13
-                    },
-                    "end": {
-                        "line": 20,
-                        "character": 13
-                    }
-                }
-            },
+            "textEditText": "completion_1(${1:null}, ${2:null}, ${3:0})",
             "command": {
                 "title": "",
                 "command": "java.completion.onDidSelect",
@@ -48,7 +24,7 @@
                 "pid": "0",
                 "rid": "0",
                 "uri": "file:///home/rgrunber/sample-projects/completion-performance/org/example/Main.java",
-                "COMPLETION_EXECUTION_TIME": "296"
+                "COMPLETION_EXECUTION_TIME": "277"
             }
         },
...
...
...
-    ]
+    ],
+    "itemDefaults": {
+        "editRange": {
+            "insert": {
+                "start": {
+                    "line": 20,
+                    "character": 13
+                },
+                "end": {
+                    "line": 20,
+                    "character": 13
+                }
+            },
+            "replace": {
+                "start": {
+                    "line": 20,
+                    "character": 13
+                },
+                "end": {
+                    "line": 20,
+                    "character": 13
+                }
+            }
+        },
+        "insertTextFormat": 2
+    }
```
</details>

`textEditText` is the right thing to use when we can't use `textEdit` because the range comes from the item defaults.

On my ~1000 completion class the overall size of the response went from 1.7MB to 984kb. So about half the size.
Calling completion about 15 times shows that there's not really any performance benefit :

**Old** : 342ms 188ms 148ms 164ms 117ms 162ms 158ms 165ms 101ms 70ms 139ms 66ms 78ms 88ms 75ms
**New**: 356ms 155ms 129ms 173ms 145ms 116ms 132ms 138ms 99ms 162ms 64ms 96ms 71ms 69ms 178ms